Add files via upload
In this project, data was collected for complete Haryana state and pollutants like pm2.5, pm 10, no2 and ozone were selected for the project. Data
was preprocessed and then was uploaded to arcgis pro and interpolation and kriging was performed.
Using Machine Learning, time series forecasting and trend analysis was done.